// src/api/back/models/TournamentBracketInfoDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    BracketMatchDto,
    BracketMatchDtoFromJSON,
    BracketMatchDtoFromJSONTyped,
    BracketMatchDtoToJSON,
    BracketParticipantDto,
    BracketParticipantDtoFromJSON,
    BracketParticipantDtoFromJSONTyped,
    BracketParticipantDtoToJSON,
    GroupDto,
    GroupDtoFromJSON,
    GroupDtoFromJSONTyped,
    GroupDtoToJSON,
    RoundDto,
    RoundDtoFromJSON,
    RoundDtoFromJSONTyped,
    RoundDtoToJSON,
    StageDto,
    StageDtoFromJSON,
    StageDtoFromJSONTyped,
    StageDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface TournamentBracketInfoDto
 */
export interface TournamentBracketInfoDto {
    /**
     * 
     * @type {Array<BracketParticipantDto>}
     * @memberof TournamentBracketInfoDto
     */
    participant: Array<BracketParticipantDto>;
    /**
     * 
     * @type {Array<StageDto>}
     * @memberof TournamentBracketInfoDto
     */
    stage: Array<StageDto>;
    /**
     * 
     * @type {Array<GroupDto>}
     * @memberof TournamentBracketInfoDto
     */
    group: Array<GroupDto>;
    /**
     * 
     * @type {Array<RoundDto>}
     * @memberof TournamentBracketInfoDto
     */
    round: Array<RoundDto>;
    /**
     * 
     * @type {Array<BracketMatchDto>}
     * @memberof TournamentBracketInfoDto
     */
    match: Array<BracketMatchDto>;
}

export function TournamentBracketInfoDtoFromJSON(json: any): TournamentBracketInfoDto {
    return TournamentBracketInfoDtoFromJSONTyped(json, false);
}

export function TournamentBracketInfoDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TournamentBracketInfoDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'participant': ((json['participant'] as Array<any>).map(BracketParticipantDtoFromJSON)),
        'stage': ((json['stage'] as Array<any>).map(StageDtoFromJSON)),
        'group': ((json['group'] as Array<any>).map(GroupDtoFromJSON)),
        'round': ((json['round'] as Array<any>).map(RoundDtoFromJSON)),
        'match': ((json['match'] as Array<any>).map(BracketMatchDtoFromJSON)),
    };
}

export function TournamentBracketInfoDtoToJSON(value?: TournamentBracketInfoDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'participant': ((value.participant as Array<any>).map(BracketParticipantDtoToJSON)),
        'stage': ((value.stage as Array<any>).map(StageDtoToJSON)),
        'group': ((value.group as Array<any>).map(GroupDtoToJSON)),
        'round': ((value.round as Array<any>).map(RoundDtoToJSON)),
        'match': ((value.match as Array<any>).map(BracketMatchDtoToJSON)),
    };
}